edit(scripts): rename files mjs to js and Game instead of chat.
