+++ client/scripts/displayStartPage.js
