+++ client/scripts/updateChat.js
